:fire: modularize object
